+++ android/src/main/java/io/sentry/RNSentryModule.java
@@ -1,4 +1,4 @@
-package io.sentry;
+package io.sentry.react;
 
 import android.content.Context;
 import android.content.pm.PackageInfo;
@@ -31,15 +31,15 @@ import java.util.logging.Logger;
 import io.sentry.android.core.AnrIntegration;
 import io.sentry.android.core.NdkIntegration;
 import io.sentry.android.core.SentryAndroid;
-import io.sentry.core.Sentry;
-import io.sentry.core.Breadcrumb;
-import io.sentry.core.Integration;
-import io.sentry.core.SentryLevel;
-import io.sentry.core.SentryOptions;
-import io.sentry.core.UncaughtExceptionHandlerIntegration;
-import io.sentry.core.protocol.SdkVersion;
-import io.sentry.core.protocol.SentryException;
-import io.sentry.core.protocol.User;
+import io.sentry.Sentry;
+import io.sentry.Breadcrumb;
+import io.sentry.Integration;
+import io.sentry.SentryLevel;
+import io.sentry.SentryOptions;
+import io.sentry.UncaughtExceptionHandlerIntegration;
+import io.sentry.protocol.SdkVersion;
+import io.sentry.protocol.SentryException;
+import io.sentry.protocol.User;
 
 @ReactModule(name = RNSentryModule.NAME)
 public class RNSentryModule extends ReactContextBaseJavaModule {